+++ resources/views/resep/index.blade.php
@@ -1,4 +1,9 @@
 @extends('layouts.main')
+
+@section('title', 'resep')
+@section('heading', 'resep')
+@section('breadcrumb', 'resep')
+
 @section('content')
 
 <a href="/createresep" class="btn btn-outline-primary">add resep</a>
@@ -26,22 +31,26 @@
         </tr>
         </thead>
         <tbody>
+
           @php
               $no = 1
           @endphp
+          
           @foreach ( $resep as $row )
             <tr>
                 <td>{{ $no++ }}</td>
+
                 <td>{{$row->Category->name ?? null}}</td>
-                <td>{{$row->judul }}</td>
-                <td> {{ $row->resep }}</td>
+                
+                <td>{{$row->judul ?? null }}</td>
+                <td> {{ $row->resep ?? null }}</td>
                 <td> 
-                  <img src="/resep_images/{{ $row->images }}" width="100" alt=""> 
+                  <img src="/resep_images/{{ $row->images ?? null }}" width="100" alt=""> 
                 </td>
                 <td>
                     <a href="/editresep/{{$row->id}}" class="btn btn-outline-warning"><i class="fa-solid fa-pen-to-square"></i></a>
 
-                    <a href="/deleteresep/{{$row->id}}" class="btn btn-outline-danger"><i class="fa-solid fa-trash"></i></a>
+                    <a href="/deleteresep/{{$row->id}}" class="btn btn-outline-danger" onclick="return confirm('delete data')" ><i class="fa-solid fa-trash"></i></a>
 
                 </td>
             </tr>
